+++ Changes.md
@@ -1,3 +1,8 @@
+# 1.1.0 - 2018-01-12
+
+- use /checkv2 endpoint (requires rspamd 1.6+)
+- support setting SMTP message from rspamd
+- support 'rewrite subject' action
 
 # 1.0.0 - 201_-__-__